+++ migrations.go
@@ -74,6 +74,7 @@ var (
 		changeNotificationTimer,
 		addContentExamples,
 		addServerIDToScheduledEvents,
+		addLoopServerToCustomRace,
 	}
 )
 
@@ -677,6 +678,37 @@ func addServerIDToScheduledEvents(s Store) error {
 
 		if err != nil {
 			return err
+		}
+	}
+
+	return nil
+}
+
+func addLoopServerToCustomRace(s Store) error {
+	logrus.Infof("Running migration: Add Loop Per Server to Custom Race")
+
+	if err := initServerID(s); err != nil {
+		return err
+	}
+
+	customRaces, err := s.ListCustomRaces()
+
+	if err != nil {
+		return err
+	}
+
+	for _, customRace := range customRaces {
+
+		if customRace.Loop {
+			customRace.LoopServer = make(map[string]bool)
+
+			customRace.LoopServer[serverID] = true
+
+			err := s.UpsertCustomRace(customRace)
+
+			if err != nil {
+				return err
+			}
 		}
 	}