+++ race_custom.go
@@ -18,11 +18,14 @@ type CustomRace struct {
 	HasCustomName, OverridePassword bool
 	ReplacementPassword             string
 
-	Created       time.Time
-	Updated       time.Time
-	Deleted       time.Time
-	UUID          uuid.UUID
-	Starred, Loop bool
+	Created time.Time
+	Updated time.Time
+	Deleted time.Time
+	UUID    uuid.UUID
+	Starred bool
+	// Deprecated: Replaced by LoopServer
+	Loop       bool
+	LoopServer map[string]bool
 
 	ForceStopTime        int
 	ForceStopWithDrivers bool
@@ -40,7 +43,11 @@ func (cr *CustomRace) GetEntryList() EntryList {
 }
 
 func (cr *CustomRace) IsLooping() bool {
-	return cr.Loop
+	if cr.LoopServer == nil {
+		return false
+	}
+
+	return cr.LoopServer[serverID]
 }
 
 func (cr *CustomRace) EventName() string {